feat: introduce "strict" mode
Passing "--strict" argument forces the hook to strictly validate a
commit message against Conventional Commits formatting. This mode is
switched off by default. It allows using fixup-like commits (starts with
 "fixup!", "amend!", and "squish!").

The motivation for adding this feature is that many people are using
"git commit --fixup" workflow during the code review process. Commits
with this prefix shouldn't be rejected because they can be sent to the
CI before being squashed.

Issue #60